pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wafer-defect-detection"
version = "0.2.1"
description = "Research-oriented wafer defect detection utilities for the WM-811K dataset."
readme = "README.md"
requires-python = ">=3.10,<3.14"
authors = [
  { name = "Anindita Paul, Brandon Pyle, Anand Rajan, Brett Rettura" },
]
dependencies = [
  "pandas>=1.3.0",
  "numpy>=1.24.0",
  "torch>=1.9.0",
  "torchvision>=0.10.0",
  "scikit-learn>=1.0.0",
  "matplotlib>=3.4.0",
  "seaborn>=0.11.0",
  "Pillow>=8.3.0",
  "opencv-python>=4.5.0",
  "scikit-image>=0.19.0",
  "tqdm>=4.60.0",
  "scipy>=1.7.0",
  "pyyaml>=5.4.0",
  "omegaconf>=2.1.0",
  "pydantic>=2.6.0",
  "joblib>=1.1.0",
  "click>=8.0.0",
  "typer>=0.12.0",
]

[project.scripts]
wafer-cli = "src.cli.main:run"

[project.optional-dependencies]
server = [
  "fastapi>=0.110.0",
  "uvicorn>=0.21.0",
  "python-multipart>=0.0.9",
]
tuning = [
  "optuna>=3.0.0",
  "optuna-dashboard>=0.13.0",
]
tuning-ray = [
  "ray>=2.0.0",
]
mlops = [
  "wandb>=0.13.0",
  "mlflow>=1.20.0",
  "tensorboard>=2.10.0",
]
dashboard = [
  "streamlit>=1.10.0",
  "plotly>=5.0.0",
]
docs = [
  "sphinx>=7.0.0",
  "sphinx-autoapi>=3.0.0",
  "sphinx-rtd-theme>=2.0.0",
  "myst-parser>=2.0.0",
  "linkify-it-py>=2.0.0",
]
notebook = [
  "jupyter>=1.0.0",
  "ipykernel>=6.0.0",
]
export = [
  "onnx>=1.12.0",
  "onnxruntime>=1.14.0",
]
research = [
  "timm>=0.9.0",
  "lightly>=1.3.0",
]
dev = [
  "pytest>=6.2.0",
  "pytest-cov>=3.0.0",
  "pytest-xdist>=2.5.0",
  "black>=21.12b0",
  "flake8>=4.0.0",
  "isort>=5.13.0",
  "mypy>=0.910",
  "fastapi>=0.110.0",
  "uvicorn>=0.21.0",
  "python-multipart>=0.0.9",
  "jupyter>=1.0.0",
  "ipykernel>=6.0.0",
  "optuna>=3.0.0",
  "optuna-dashboard>=0.13.0",
  "wandb>=0.13.0",
  "mlflow>=1.20.0",
  "streamlit>=1.10.0",
  "plotly>=5.0.0",
  "sphinx>=4.3.0",
  "sphinx-rtd-theme>=1.0.0",
  "onnx>=1.12.0",
  "onnxruntime>=1.14.0",
  "tensorboard>=2.10.0",
  "timm>=0.9.0",
  "lightly>=1.3.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
